requirements.txt
streamlit==1.22.0
boto3==1.26.142
python-dotenv==1.0.0
